Feature docker (#1)
* Add docker-compose.yml

* Add docker-compose.yml

* Add relation to entities and refactor

+++ src/main/java/pl/burbaniak/classicdatabase/dto/CustomerDto.java
@@ -19,7 +19,7 @@ public class CustomerDto {
     private String state;
     private String postalCode;
     private String country;
-    private int salesRepEmployeeNumber;
+    private Long salesRepEmployeeNumber;
     private double creditLimit;
 
     public static CustomerDto from(Customer customer){
